fix: 베스트 반찬 API endpoint 수정
- RESTful하도록 반찬 목록 조회에 대해 /products를 추가
- 해당 API의 컨트롤러 위치 변경
- 테스트 수정
