+++ src/main/java/codesquad/category/web/ApiCategoryController.java
@@ -2,17 +2,12 @@ package codesquad.category.web;
 
 import codesquad.RestResponse;
 import codesquad.category.domain.*;
-import codesquad.product.domain.Product;
 import codesquad.security.HttpSessionUtils;
 import codesquad.user.domain.User;
 import codesquad.user.domain.UserRepository;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.annotations.Cache;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -51,15 +46,6 @@ public class ApiCategoryController {
     public ResponseEntity<RestResponse> getBestCategory() {
         List<BestCategory> bestCategories = bestCategoryRepository.findAll();
         return ResponseEntity.ok(RestResponse.of(bestCategories));
-    }
-
-    @ApiOperation(value = "게시판 등록")
-    @ApiImplicitParam(name = "id", value = "이벤트 반찬 카테고리 id", required = true, dataType = "long")
-    @GetMapping("/category/best/{id}")
-    @Cacheable(value="bestCategoryProducts")
-    public ResponseEntity<RestResponse> getProductList(@PathVariable Long id) {
-        BestCategory bestCategory = bestCategoryRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-        return ResponseEntity.ok(RestResponse.of(bestCategory.getProducts()));
     }
 
     @PostMapping("/admin/category")
